Partial implementation during discussion with Livia. WILL BREAK TESTS! (Livia to fix)
